+++ command.h
@@ -7,6 +7,8 @@
 #ifndef COMMAND_H
 #define COMMAND_H
 
+#include <stdint.h>
+
 typedef enum {
     COORDINATE,
     START,
@@ -15,11 +17,11 @@ typedef enum {
 } command_type;
 
 typedef struct {
-    int x;
-    int y;
+    uint16_t x;
+    uint16_t y;
     command_type type; //1 is start, 2 is lift pen, 3 is stop
 } command_t;
 
-command_t* create_command(int x, int y, command_type type);
+command_t* create_command(uint16_t x, uint16_t y, command_type type);
 
 #endif
